cegs_gnn/requirements.txt
torch>=1.9.0
pytorch-geometric>=2.0.0
numpy>=1.20.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
networkx>=2.6.0
tqdm>=4.62.0
pandas>=1.3.0
jupyter>=1.0.0 